requirements.txt
functions-framework==3.*
pandas==2.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
openpyxl==3.*
